requirements.txt
Django==2.1.7
django-simple-history==2.7.0
numpy==1.16.2
pandas==0.24.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
